chore: rename apiv2 app.ts to bootstrap.ts
